refactor total volume data type

+++ feature_coin_list/src/main/java/com/ryankoech/krypto/feature_coin_list/domain/entity/Coin.kt
@@ -11,5 +11,5 @@ data class Coin(
     val marketCap : Long,
     val allTimeHigh : Double,
     val high24Hr : Double,
-    val totalVolume : Long,
+    val totalVolume : Double,
 )
